+++ valens/database.py
@@ -1,11 +1,13 @@
 import sqlite3
+from datetime import datetime
 from pathlib import Path
+from shutil import copy
+from time import sleep
 
 from alembic import command, runtime, script
 from alembic.config import Config
 from flask import current_app, g
-from sqlalchemy import create_engine, event, inspect, pool
-from sqlalchemy.engine import Engine
+from sqlalchemy import Connection, Engine, create_engine, event, inspect, pool
 from sqlalchemy.orm import Session, scoped_session, sessionmaker
 from werkzeug.local import LocalProxy
 
@@ -15,19 +17,21 @@ alembic_cfg = Config()
 alembic_cfg.set_main_option("script_location", "valens:migrations")
 
 
-@event.listens_for(Engine, "connect")
-def _set_sqlite_pragma(
-    dbapi_connection: sqlite3.Connection, _: pool.base._ConnectionRecord
-) -> None:
-    if current_app.config["SQLITE_FOREIGN_KEY_SUPPORT"]:
-        cursor = dbapi_connection.cursor()
-        cursor.execute("PRAGMA foreign_keys=ON")
-        cursor.close()
+def db_file() -> Path:
+    return Path(current_app.config["DATABASE"].removeprefix("sqlite:///"))
+
+
+def db_dir() -> Path:
+    return db_file().parent
+
+
+def upgrade_lock_file() -> Path:
+    return db_dir() / "valens_upgrade.lock"
 
 
 def get_engine() -> Engine:
     config.check_app_config()
-    Path(current_app.config["DATABASE"].split(":")[1]).parent.mkdir(exist_ok=True)
+    db_dir().mkdir(exist_ok=True)
     return create_engine(current_app.config["DATABASE"])
 
 
@@ -40,8 +44,10 @@ def get_scoped_session() -> scoped_session[Session]:
 def get_session() -> Session:
     if "db_session" not in g:
         if not inspect(get_engine()).get_table_names():
-            init_db()
+            init()
         g.db_session = get_scoped_session()()
+
+    _upgrade(g.db_session.connection())
 
     return g.db_session
 
@@ -53,7 +59,7 @@ def remove_session() -> None:
     get_scoped_session().remove()
 
 
-def init_db() -> None:
+def init() -> None:
     print("Creating database")
 
     models.Base.query = get_scoped_session().query_property()
@@ -62,14 +68,42 @@ def init_db() -> None:
     command.stamp(alembic_cfg, "head")
 
 
-def upgrade_db() -> None:
-    current = runtime.migration.MigrationContext.configure(
-        get_session().connection()
-    ).get_current_revision()
+def upgrade() -> None:
+    get_session()
+
+
+def _upgrade(connection: Connection) -> None:
+    current = runtime.migration.MigrationContext.configure(connection).get_current_revision()
     head = script.ScriptDirectory.from_config(alembic_cfg).get_current_head()
 
     if current != head:
-        print(f"Upgrading database from {current} to {head}")
-        command.upgrade(alembic_cfg, "head")
-    else:
-        print("No upgrade necessary")
+        try:
+            upgrade_lock_file().touch(exist_ok=False)
+
+            print(f"Upgrading database from {current} to {head}")
+
+            copy(
+                db_file(),
+                db_file().with_suffix(
+                    f".db.backup_{current}_{datetime.now().isoformat(timespec='seconds')}"
+                ),
+            )
+            command.upgrade(alembic_cfg, "head")
+
+            upgrade_lock_file().unlink()
+
+        except FileExistsError:
+            print("Waiting for completion of database upgrade")
+
+            while upgrade_lock_file().exists():
+                sleep(1)
+
+
+@event.listens_for(Engine, "connect")
+def _set_sqlite_pragma(
+    dbapi_connection: sqlite3.Connection, _: pool.base._ConnectionRecord
+) -> None:
+    if current_app.config["SQLITE_FOREIGN_KEY_SUPPORT"]:
+        cursor = dbapi_connection.cursor()
+        cursor.execute("PRAGMA foreign_keys=ON")
+        cursor.close()